minor
Former-commit-id: db43904c92efb38acda471448b8129f641615fd6

+++ spimagine/data_model.py
@@ -238,7 +238,7 @@ class DemoData(GenericData):
         return self.nT
 
     def __getitem__(self,pos):
-        return self.data*np.exp(-.3*pos)
+        return (self.data*np.exp(-.3*pos)).astype(np.uint16)
 
 
 class EmptyData(GenericData):
